+++ SpringScms/SpringSCMS/src/main/java/com/info/Service/StudentService.java
@@ -35,7 +35,7 @@ public class StudentService {
 		repo.updatePoints(studPoints, studentId);
 	}
 	
-	public void updateWallet(int fee,int studentId)
+	public void updateWallet(double fee,int studentId)
 	{
 		repo.updateWallet(fee, studentId);
 	}
